+++ Engine_SOURCE/yaPlayScene.h
@@ -3,6 +3,7 @@
 
 namespace ya
 {
+	class MeshData;
 	class PlayScene : public Scene
 	{
 	public:
@@ -17,6 +18,11 @@ namespace ya
 		virtual void OnEnter() override;
 		virtual void OnExit() override;
 
+
+		void Test() { int a = 0; }
 	private:
+		//for test
+		std::shared_ptr<MeshData> mMeshData;
+
 	};
 }
