+++ Engine_SOURCE/yaTitleScene.h
@@ -3,7 +3,6 @@
 
 namespace ya
 {
-	class MeshData;
 	class TitleScene : public Scene
 	{
 	public:
@@ -18,11 +17,6 @@ namespace ya
 		virtual void OnEnter() override;
 		virtual void OnExit() override;
 
-
-		void Test() { int a = 0; }
 	private:
-		//for test
-		std::shared_ptr<MeshData> mMeshData;
-
 	};
 }
